+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## Version 1.16.0
 
-Released on ?.
+Released on 2025-04-10.
 
 ✨ New Features / Improvements